Move `ToValType` to `convert`
