+++ ext/src/ruby_api/convert.rs
@@ -1,8 +1,18 @@
-use crate::{err, error};
-use magnus::{Error, TypedData, Value};
+use crate::{define_rb_intern, err, error};
+use magnus::{Error, Symbol, TypedData, Value};
 use wasmtime::{ExternRef, Val, ValType};
 
 use super::{func::Func, memory::Memory, store::StoreContextValue};
+
+define_rb_intern!(
+    I32 => "i32",
+    I64 => "i64",
+    F32 => "f32",
+    F64 => "f64",
+    V128 => "v128",
+    FUNCREF => "funcref",
+    EXTERNREF => "externref",
+);
 
 pub trait ToRubyValue {
     fn to_ruby_value(&self, store: &StoreContextValue) -> Result<Value, Error>;
@@ -81,6 +91,60 @@ impl ToExtern for Value {
                 format!("unexpected extern: {}", self.inspect()),
             ))
         }
+    }
+}
+
+pub trait ToSym {
+    fn to_sym(self) -> Symbol;
+}
+
+impl ToSym for ValType {
+    fn to_sym(self) -> Symbol {
+        match self {
+            ValType::I32 => Symbol::from(*I32),
+            ValType::I64 => Symbol::from(*I64),
+            ValType::F32 => Symbol::from(*F32),
+            ValType::F64 => Symbol::from(*F64),
+            ValType::V128 => Symbol::from(*V128),
+            ValType::FuncRef => Symbol::from(*FUNCREF),
+            ValType::ExternRef => Symbol::from(*EXTERNREF),
+        }
+    }
+}
+pub trait ToValType {
+    fn to_val_type(&self) -> Result<ValType, Error>;
+}
+
+impl ToValType for Value {
+    fn to_val_type(&self) -> Result<ValType, Error> {
+        if let Ok(symbol) = self.try_convert::<Symbol>() {
+            if let Ok(true) = symbol.equal(Symbol::from(*I32)) {
+                return Ok(ValType::I32);
+            }
+            if let Ok(true) = symbol.equal(Symbol::from(*I64)) {
+                return Ok(ValType::I64);
+            }
+            if let Ok(true) = symbol.equal(Symbol::from(*F32)) {
+                return Ok(ValType::F32);
+            }
+            if let Ok(true) = symbol.equal(Symbol::from(*F64)) {
+                return Ok(ValType::F64);
+            }
+            if let Ok(true) = symbol.equal(Symbol::from(*V128)) {
+                return Ok(ValType::V128);
+            }
+            if let Ok(true) = symbol.equal(Symbol::from(*FUNCREF)) {
+                return Ok(ValType::FuncRef);
+            }
+            if let Ok(true) = symbol.equal(Symbol::from(*EXTERNREF)) {
+                return Ok(ValType::ExternRef);
+            }
+        }
+
+        err!(
+            "invalid Webassembly type, expected one of [:i32, :i64, :f32, :f64, :v128, :funcref, :externref], got {:}",
+            self.inspect()
+        )
     }
 }
